Move quiz/slack to slack
